Try another Travis script
This easier-to-read script uses Cabal for builds. I don't use Cabal, but for
CI it may be better because it allows to build in more adverse environment
and what builds with Cabal will sure build with Stack and appropriate
resolver. I hope this should help catch things like #46 in the future.

This also adds a check that we can generate tarball distribution without
issues and that Haddock coverage is 100%.