Update LTS to 12.0
Updated LTS to 12.0. In order to do this we needed to take care of
validity, which has a new API in 12.0. As a result, tests are updated.
Changes are almost 100% superficial, with the only semantical change
being added shrinking for unchecked generators, although they are
1) Never used in the current code.
2) Just wrappers around String shrinking.

As a result we no longer have validity as an external dependency, as it
is in LTS 12.0.