Add MonadError in addition to MonadThrow.
This abstracts Include.hs to be able to swap between `MonadError` and
`MonadThrow`. The `MonadError` implementations are exposed via
`Path.Except[.*]`, while `MonadThrow` remains in `Path[.*]`.

Some shortcomings and considerations:
- there is now a dependency on MTL
- `-Wunused-imports` is disabled so both classes can be imported
  without having to conditionalize anything
- there are no tests run against `Path.Except[.*]` yet

A slightly different approach could be to publish two packages -- `path`
and `path-except`. They would expose the exact same modules (`.Except`
would disappear), but one would have a dependency on `exceptions` and
the other on `mtl`. It's unlikely that anyone would want to use both the
MonadThrow and MonadError approaches in the same project, so this could
reduce their dependency footprint, but it does mean I have to fix the
unused imports.

Fixes #149.